+++ tests/fixtures/join/interface_body/extends.in.ts
@@ -1,0 +1,4 @@
+interface Foo extends Bar {
+  a: string;
+  b: number;
+}