+++ tests/fixtures/split/interface_body/extends.out.ts
@@ -1,0 +1,4 @@
+interface Foo extends Bar, Baz {
+  a: string;
+  b: number;
+}