+++ tests/fixtures/split/interface_body/index-signature.out.ts
@@ -1,0 +1,4 @@
+interface Foo {
+  [key: string]: number;
+  a: number;
+}